DJVM fixes for public key caching
