+++ core-deterministic/build.gradle
@@ -82,6 +82,7 @@ def patchCore = tasks.register('patchCore', Zip) {
         exclude 'net/corda/core/serialization/internal/CheckpointSerializationFactory*.class'
         exclude 'net/corda/core/internal/rules/*.class'
         exclude 'net/corda/core/internal/utilities/PrivateInterner*.class'
+        exclude 'net/corda/core/crypto/internal/PublicKeyCache*.class'
     }
 
     reproducibleFileOrder = true
