+++ core-deterministic/src/main/kotlin/net/corda/core/crypto/internal/PublicKeyCache.kt
@@ -1,0 +1,21 @@
+package net.corda.core.crypto.internal
+
+import net.corda.core.utilities.ByteSequence
+import java.security.PublicKey
+
+@Suppress("unused")
+object PublicKeyCache {
+    @Suppress("UNUSED_PARAMETER")
+    fun bytesForCachedPublicKey(key: PublicKey): ByteSequence? {
+        return null
+    }
+
+    @Suppress("UNUSED_PARAMETER")
+    fun publicKeyForCachedBytes(bytes: ByteSequence): PublicKey? {
+        return null
+    }
+
+    fun cachePublicKey(key: PublicKey): PublicKey {
+        return key
+    }
+}
